refactor(protocol): SEC-003 리뷰 반영 — size_t 비교 + over-claim 테스트
PR #30 리뷰(Claude/Gemini) 반영:
- frame.c: header.length 교차검증을 uint16_t 캐스트 대신 size_t 비교로 —
  hdr length를 size_t로 승격해 truncating 캐스트 제거(프레임은 이미
  OPC_FRAME_MAX로 상한이라 실제 truncation은 없었으나 의존성 제거).
- test_codec: length over-claim(64B 프레임에 length=100) 양성 테스트 추가 —
  기존 under-claim(length=0)과 함께 양방향 불일치 거부를 검증.

make check 전체 통과(PASS 290 / FAIL 0), arm64 PLATFORM=nxp 크로스빌드 확인.

+++ protocol/frame.c
@@ -81,10 +81,11 @@ int opc_frame_parse(const uint8_t *frame, size_t frame_len,
      * whose length disagrees with its actual size is malformed, so reject it
      * here rather than trusting a lying length downstream. frame_len >= the
      * fixed header is already guaranteed above, so the subtraction cannot
-     * underflow, and (1424 - 8) fits a uint16_t. Self-consistent short-body
-     * frames still pass this gate and reach the per-command unpack, which
-     * issues OPC_ERR_PACKET_SIZE as before. */
-    if (hdr_out->length != (uint16_t)(frame_len - OPC_FIXED_HEADER_SIZE)) {
+     * underflow. The comparison is done in size_t (hdr length promoted) so no
+     * truncating cast is needed — both over- and under-claims are rejected.
+     * Self-consistent short-body frames still pass this gate and reach the
+     * per-command unpack, which issues OPC_ERR_PACKET_SIZE as before. */
+    if ((size_t)hdr_out->length != frame_len - OPC_FIXED_HEADER_SIZE) {
         return -1;
     }
     /* Body present only once the frame reaches the full common header.
